Renamed picker as not using accordion
